+++ packages/data_tables/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.3.0 - 04.06.2019
 
-- Making Example Desktop Aware
+* Making Example Desktop Aware
 
 ## 0.2.4